Refactor httpSearch service into its own module Reorganizing the code this way facilitiates using different inventory stores for dev vs unit testing.
